backend/requirements.txt
django
djangorestframework
environs
django-phonenumber-field[phonenumbers]
phonenumbers
psycopg2
pillow